add saveAll

+++ lab2/src/main/java/ru/gaidamaka/service/OsmPersistService.java
@@ -6,6 +6,6 @@ public interface OsmPersistService {
 
     void persist(Osm osm);
 
-    void persist(Osm osm, long limit);
+    void persist(Osm osm, int limit);
 
 }
